+++ packages/site/src/components/ContentLoader.tsx
@@ -22,6 +22,9 @@ export const ContentLoader = () => {
     case location.pathname.startsWith("/hooks"):
       type = "hooks";
       break;
+    case location.pathname.startsWith("/guides"):
+      type = "guides";
+      break;
     default:
       type = "content";
   }