+++ packages/site/src/maps/index.ts
@@ -2,6 +2,7 @@ import { designContentMap } from "./designContent";
 import { contentContentMap } from "./contentContent";
 import { changelogContentMap } from "./changelogContent";
 import { hooksContentMap } from "./hooksContent";
+import { guidesContentMap } from "./guidesContent";
 import { ContentMap } from "../types/maps";
 
 export const contentMap: ContentMap = {
@@ -9,4 +10,5 @@ export const contentMap: ContentMap = {
   design: designContentMap,
   hooks: hooksContentMap,
   changelog: changelogContentMap,
+  guides: guidesContentMap,
 };